(refactor) clean up unnecessary types and move conversion functions to models.go
